+++ R/mod_main_panel.R
@@ -26,7 +26,7 @@ mod_main_panel_ui <- function(id) {
       inputId = ns("sensor_select"),
       label = tags$h4("Sensor"), 
       selected = NULL,
-      choices = list("Loading Sensors..." = NULL)
+      choices = list("Loading Sensors..." = NULL), 
     ),
 
     tags$hr(), 
@@ -68,13 +68,6 @@ mod_main_panel_ui <- function(id) {
 
     ),
     
-    # Hacky way to get the client timezone stored in the client object
-    # shinyjs::hidden(textInput(
-    #   inputId = ns("client_tz"),
-    #   "client time",
-    #   value = ""
-    # )),
-    
     tags$hr(), 
     
     fluidRow(
@@ -92,8 +85,6 @@ mod_main_panel_ui <- function(id) {
         width = 7
       ),
       column(
-        bs_embed_tooltip(
-          title = "Copy URL",
           actionLink(
             inputId = ns("share_button"), 
             label = tags$div(HTML('<svg width="1em" height="1em" viewBox="0 0 16 16" class="bi bi-link-45deg" fill="currentColor" xmlns="http://www.w3.org/2000/svg">
@@ -102,24 +93,10 @@ mod_main_panel_ui <- function(id) {
   <path d="M6.586 4.672A3 3 0 0 0 7.414 9.5l.775-.776a2 2 0 0 1-.896-3.346L9.12 3.55a2 2 0 0 1 2.83 2.83l-.793.792c.112.42.155.855.128 1.287l1.372-1.372a3 3 0 0 0-4.243-4.243L6.586 4.672z"/>
   <path d="M10 9.5a2.99 2.99 0 0 0 .288-1.46l-.167.167a1.99 1.99 0 0 1-.896.518 1.99 1.99 0 0 1-.518.896l-.167.167A3.004 3.004 0 0 0 10 9.501z"/>
 </svg> Share...'))
-          )
-        ),
-        
-        # Add neat interactivity to the copy tooltip onclick:copied!
+          ) %>% bs_embed_tooltip(
+            title = "Get URL"
+          ),
         tags$style("#main_panel_ui_1-share_button { text-align:center; }"),
-        tags$script(
-          HTML(
-            '$("#main_panel_ui_1-share_button").on("click", () => {
-            $("#main_panel_ui_1-share_button")
-              .attr("data-original-title", "Copied!")
-              .tooltip("fixTitle").tooltip("show");
-            });
-            $("#main_panel_ui_1-share_button").on("hidden.bs.tooltip", () => {
-              $("#main_panel_ui_1-share_button")
-                .attr("data-original-title", "Copy URL")
-                .tooltip("fixTitle");
-            });')
-        ),
         width =  5
       ), 
       
@@ -199,11 +176,11 @@ mod_main_panel_server <- function(input, output, session, usr) {
         with({
           # Check diff bewteen sensors aobj in sensor obj and pas obj and only use
           # the sensors with mutual existence
-          pas_communities <- na.omit(unique(id2com(pas[['communityRegion']])))
-          pas_labels <- na.omit(unique(pas[['label']]))
+          pas_communities <- na.omit(unique(id2com(pas$communityRegion)))
+          pas_labels <- na.omit(unique(pas$label))
           
-          sensors_communities <- na.omit(unique(id2com(sensors[["meta"]][['communityRegion']])))
-          sensors_labels <- na.omit(unique(sensors[["meta"]][['label']]))
+          sensors_communities <- na.omit(unique(id2com(sensors$meta$communityRegion)))
+          sensors_labels <- na.omit(unique(sensors$meta$label))
           
           community_choices <- sensors_communities[sensors_communities %in% pas_communities]
           sensor_choices <- sensors_labels[sensors_labels %in% pas_labels]
@@ -264,9 +241,6 @@ mod_main_panel_server <- function(input, output, session, usr) {
       
       yr <- lubridate::year(input$date_select)
       
-      # Plot down to avoid weird bugs 
-      plotDown()
-      
       # update the client object date selections
       usr$selected$sd <- sd
       usr$selected$ed <- ed
@@ -324,19 +298,14 @@ mod_main_panel_server <- function(input, output, session, usr) {
   
   # Write the url to the user clipboard on share click
   observeEvent(
-    ignoreNULL = TRUE, 
-    ignoreInit = TRUE, 
     eventExpr = { input$share_button }, 
     handlerExpr = {
       url <- usr$url
       
       tryCatch(
         expr = {
-          # Why does this break shiny? 
-          # A: see allow_non_interactive in docs
-          #write_clip(url, allow_non_interactive = TRUE)
-          shinyjs::runjs('
-                         ')
+          # Just show a modal with the URL - copy and pasting is actually complicated..
+          showBookmarkUrlModal(usr$selected$url)
         }, 
         error = function(err) {
           logger.error(err)
@@ -345,6 +314,7 @@ mod_main_panel_server <- function(input, output, session, usr) {
       )
     }
   )
+  
   
   # Watch the current page and tab. 
   # if on the latest page, hide the date range input